Ease the transfer to website and more. See below.
CHG: First name of "Mitch Marner" is now "Mitch", not "Mitchell".

ADD: Implement an alternative method to retrieve information from
OfficePools site. It's not more manual. This is because we can't
download stuff from their page for the moment.

ADD: When there is a difference between OfficePools and NHl for a
player, we may now continue to check all participants instead of
stopping at the first one like before.

ADD: Will now zip all files to upload to website and unzip once there.

ADD: Now generate a .html file with absolute links so we can copy and
paste ranking grids to an email and send to colleagues.

CHG: Version is not 1.0.3 instead of 0.0.3.

ADD: Include in the repository the image of bluberi man.